+++ cloudsplaining/shared/constants.py
@@ -6,6 +6,7 @@ from cloudsplaining.shared.validation import check_exclusions_schema
 
 logger = logging.getLogger(__name__)
 
+
 PACKAGE_DIR = str(
     os.path.abspath(os.path.join(os.path.dirname(__file__), os.path.pardir))
 )