+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 * Removed "Stats for Nerds"
+* Cleaned up report some more
+* Made Pylint happy
 
 ## 0.0.1 (2020-04-23)
 * Improved HTML report. Includes scoring.